Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/css/css-transforms/animation/transform-skew-composition-expected.txt


FAIL Compositing: property <transform> underlying [skewX(10deg)] from add [skewX(30deg)] to add [skewX(50deg)] at (-0.5) should be [skewX(10deg) skewX(20deg)] assert_equals: expected "matrix ( 1 , 0 , 0.54 , 1 , 0 , 0 ) " but got "matrix ( 1 , 0 , 0.36 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skewX(10deg)] from add [skewX(30deg)] to add [skewX(50deg)] at (0) should be [skewX(10deg) skewX(30deg)] assert_equals: expected "matrix ( 1 , 0 , 0.75 , 1 , 0 , 0 ) " but got "matrix ( 1 , 0 , 0.58 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skewX(10deg)] from add [skewX(30deg)] to add [skewX(50deg)] at (0.25) should be [skewX(10deg) skewX(35deg)] assert_equals: expected "matrix ( 1 , 0 , 0.88 , 1 , 0 , 0 ) " but got "matrix ( 1 , 0 , 0.7 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skewX(10deg)] from add [skewX(30deg)] to add [skewX(50deg)] at (0.5) should be [skewX(10deg) skewX(40deg)] assert_equals: expected "matrix ( 1 , 0 , 1.02 , 1 , 0 , 0 ) " but got "matrix ( 1 , 0 , 0.84 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skewX(10deg)] from add [skewX(30deg)] to add [skewX(50deg)] at (0.75) should be [skewX(10deg) skewX(45deg)] assert_equals: expected "matrix ( 1 , 0 , 1.18 , 1 , 0 , 0 ) " but got "matrix ( 1 , 0 , 1 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skewX(10deg)] from add [skewX(30deg)] to add [skewX(50deg)] at (1) should be [skewX(10deg) skewX(50deg)] assert_equals: expected "matrix ( 1 , 0 , 1.37 , 1 , 0 , 0 ) " but got "matrix ( 1 , 0 , 1.19 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skewX(10deg)] from add [skewX(30deg)] to add [skewX(50deg)] at (1.5) should be [skewX(10deg) skewX(60deg)] assert_equals: expected "matrix ( 1 , 0 , 1.91 , 1 , 0 , 0 ) " but got "matrix ( 1 , 0 , 1.73 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skewY(10deg)] from add [skewY(30deg)] to add [skewY(50deg)] at (-0.5) should be [skewY(10deg) skewY(20deg)] assert_equals: expected "matrix ( 1 , 0.54 , 0 , 1 , 0 , 0 ) " but got "matrix ( 1 , 0.36 , 0 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skewY(10deg)] from add [skewY(30deg)] to add [skewY(50deg)] at (0) should be [skewY(10deg) skewY(30deg)] assert_equals: expected "matrix ( 1 , 0.75 , 0 , 1 , 0 , 0 ) " but got "matrix ( 1 , 0.58 , 0 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skewY(10deg)] from add [skewY(30deg)] to add [skewY(50deg)] at (0.25) should be [skewY(10deg) skewY(35deg)] assert_equals: expected "matrix ( 1 , 0.88 , 0 , 1 , 0 , 0 ) " but got "matrix ( 1 , 0.7 , 0 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skewY(10deg)] from add [skewY(30deg)] to add [skewY(50deg)] at (0.5) should be [skewY(10deg) skewY(40deg)] assert_equals: expected "matrix ( 1 , 1.02 , 0 , 1 , 0 , 0 ) " but got "matrix ( 1 , 0.84 , 0 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skewY(10deg)] from add [skewY(30deg)] to add [skewY(50deg)] at (0.75) should be [skewY(10deg) skewY(45deg)] assert_equals: expected "matrix ( 1 , 1.18 , 0 , 1 , 0 , 0 ) " but got "matrix ( 1 , 1 , 0 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skewY(10deg)] from add [skewY(30deg)] to add [skewY(50deg)] at (1) should be [skewY(10deg) skewY(50deg)] assert_equals: expected "matrix ( 1 , 1.37 , 0 , 1 , 0 , 0 ) " but got "matrix ( 1 , 1.19 , 0 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skewY(10deg)] from add [skewY(30deg)] to add [skewY(50deg)] at (1.5) should be [skewY(10deg) skewY(60deg)] assert_equals: expected "matrix ( 1 , 1.91 , 0 , 1 , 0 , 0 ) " but got "matrix ( 1 , 1.73 , 0 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skew(10deg, 20deg)] from add [skew(30deg, 10deg)] to add [skew(50deg, 50deg)] at (-0.5) should be [skew(10deg, 20deg) skew(20deg, -10deg)] assert_equals: expected "matrix ( 0.97 , 0.19 , 0.54 , 1.13 , 0 , 0 ) " but got "matrix ( 1 , - 0.18 , 0.36 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skew(10deg, 20deg)] from add [skew(30deg, 10deg)] to add [skew(50deg, 50deg)] at (0) should be [skew(10deg, 20deg) skew(30deg, 10deg)] assert_equals: expected "matrix ( 1.03 , 0.54 , 0.75 , 1.21 , 0 , 0 ) " but got "matrix ( 1 , 0.18 , 0.58 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skew(10deg, 20deg)] from add [skew(30deg, 10deg)] to add [skew(50deg, 50deg)] at (0.25) should be [skew(10deg, 20deg) skew(35deg, 20deg)] assert_equals: expected "matrix ( 1.06 , 0.73 , 0.88 , 1.25 , 0 , 0 ) " but got "matrix ( 1 , 0.36 , 0.7 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skew(10deg, 20deg)] from add [skew(30deg, 10deg)] to add [skew(50deg, 50deg)] at (0.5) should be [skew(10deg, 20deg) skew(40deg, 30deg)] assert_equals: expected "matrix ( 1.1 , 0.94 , 1.02 , 1.31 , 0 , 0 ) " but got "matrix ( 1 , 0.58 , 0.84 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skew(10deg, 20deg)] from add [skew(30deg, 10deg)] to add [skew(50deg, 50deg)] at (0.75) should be [skew(10deg, 20deg) skew(45deg, 40deg)] assert_equals: expected "matrix ( 1.15 , 1.2 , 1.18 , 1.36 , 0 , 0 ) " but got "matrix ( 1 , 0.84 , 1 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skew(10deg, 20deg)] from add [skew(30deg, 10deg)] to add [skew(50deg, 50deg)] at (1) should be [skew(10deg, 20deg) skew(50deg, 50deg)] assert_equals: expected "matrix ( 1.21 , 1.56 , 1.37 , 1.43 , 0 , 0 ) " but got "matrix ( 1 , 1.19 , 1.19 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skew(10deg, 20deg)] from add [skew(30deg, 10deg)] to add [skew(50deg, 50deg)] at (1.5) should be [skew(10deg, 20deg) skew(60deg, 70deg)] assert_equals: expected "matrix ( 1.48 , 3.11 , 1.91 , 1.63 , 0 , 0 ) " but got "matrix ( 1 , 2.75 , 1.73 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skewX(45deg)] from accumulate [skewX(30deg)] to accumulate [skewX(70deg)] at (-0.5) should be [skewX(55deg)] assert_equals: expected "matrix ( 1 , 0 , 1.43 , 1 , 0 , 0 ) " but got "matrix ( 1 , 0 , 0.18 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skewX(45deg)] from accumulate [skewX(30deg)] to accumulate [skewX(70deg)] at (0) should be [skewX(75deg)] assert_equals: expected "matrix ( 1 , 0 , 3.73 , 1 , 0 , 0 ) " but got "matrix ( 1 , 0 , 0.58 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skewX(45deg)] from accumulate [skewX(30deg)] to accumulate [skewX(70deg)] at (0.25) should be [skewX(85deg)] assert_equals: expected "matrix ( 1 , 0 , 11.43 , 1 , 0 , 0 ) " but got "matrix ( 1 , 0 , 0.84 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skewX(45deg)] from accumulate [skewX(30deg)] to accumulate [skewX(70deg)] at (0.5) should be [skewX(95deg)] assert_equals: expected "matrix ( 1 , 0 , - 11.43 , 1 , 0 , 0 ) " but got "matrix ( 1 , 0 , 1.19 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skewX(45deg)] from accumulate [skewX(30deg)] to accumulate [skewX(70deg)] at (0.75) should be [skewX(105deg)] assert_equals: expected "matrix ( 1 , 0 , - 3.73 , 1 , 0 , 0 ) " but got "matrix ( 1 , 0 , 1.73 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skewX(45deg)] from accumulate [skewX(30deg)] to accumulate [skewX(70deg)] at (1) should be [skewX(115deg)] assert_equals: expected "matrix ( 1 , 0 , - 2.14 , 1 , 0 , 0 ) " but got "matrix ( 1 , 0 , 2.75 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skewX(45deg)] from accumulate [skewX(30deg)] to accumulate [skewX(70deg)] at (1.5) should be [skewX(135deg)] assert_equals: expected "matrix ( 1 , 0 , - 1 , 1 , 0 , 0 ) " but got "matrix ( 1 , 0 , 16331239353195370 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skewY(45deg)] from accumulate [skewY(30deg)] to accumulate [skewY(70deg)] at (-0.5) should be [skewY(55deg)] assert_equals: expected "matrix ( 1 , 1.43 , 0 , 1 , 0 , 0 ) " but got "matrix ( 1 , 0.18 , 0 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skewY(45deg)] from accumulate [skewY(30deg)] to accumulate [skewY(70deg)] at (0) should be [skewY(75deg)] assert_equals: expected "matrix ( 1 , 3.73 , 0 , 1 , 0 , 0 ) " but got "matrix ( 1 , 0.58 , 0 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skewY(45deg)] from accumulate [skewY(30deg)] to accumulate [skewY(70deg)] at (0.25) should be [skewY(85deg)] assert_equals: expected "matrix ( 1 , 11.43 , 0 , 1 , 0 , 0 ) " but got "matrix ( 1 , 0.84 , 0 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skewY(45deg)] from accumulate [skewY(30deg)] to accumulate [skewY(70deg)] at (0.5) should be [skewY(95deg)] assert_equals: expected "matrix ( 1 , - 11.43 , 0 , 1 , 0 , 0 ) " but got "matrix ( 1 , 1.19 , 0 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skewY(45deg)] from accumulate [skewY(30deg)] to accumulate [skewY(70deg)] at (0.75) should be [skewY(105deg)] assert_equals: expected "matrix ( 1 , - 3.73 , 0 , 1 , 0 , 0 ) " but got "matrix ( 1 , 1.73 , 0 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skewY(45deg)] from accumulate [skewY(30deg)] to accumulate [skewY(70deg)] at (1) should be [skewY(115deg)] assert_equals: expected "matrix ( 1 , - 2.14 , 0 , 1 , 0 , 0 ) " but got "matrix ( 1 , 2.75 , 0 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skewY(45deg)] from accumulate [skewY(30deg)] to accumulate [skewY(70deg)] at (1.5) should be [skewY(135deg)] assert_equals: expected "matrix ( 1 , - 1 , 0 , 1 , 0 , 0 ) " but got "matrix ( 1 , 16331239353195370 , 0 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skew(10deg, 45deg)] from accumulate [skew(20deg, 30deg)] to accumulate [skew(40deg, 70deg)] at (-0.5) should be [skew(20deg, 55deg)] assert_equals: expected "matrix ( 1 , 1.43 , 0.36 , 1 , 0 , 0 ) " but got "matrix ( 1 , 0.18 , 0.18 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skew(10deg, 45deg)] from accumulate [skew(20deg, 30deg)] to accumulate [skew(40deg, 70deg)] at (0) should be [skew(30deg, 75deg)] assert_equals: expected "matrix ( 1 , 3.73 , 0.58 , 1 , 0 , 0 ) " but got "matrix ( 1 , 0.58 , 0.36 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skew(10deg, 45deg)] from accumulate [skew(20deg, 30deg)] to accumulate [skew(40deg, 70deg)] at (0.25) should be [skew(35deg, 85deg)] assert_equals: expected "matrix ( 1 , 11.43 , 0.7 , 1 , 0 , 0 ) " but got "matrix ( 1 , 0.84 , 0.47 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skew(10deg, 45deg)] from accumulate [skew(20deg, 30deg)] to accumulate [skew(40deg, 70deg)] at (0.5) should be [skew(40deg, 95deg)] assert_equals: expected "matrix ( 1 , - 11.43 , 0.84 , 1 , 0 , 0 ) " but got "matrix ( 1 , 1.19 , 0.58 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skew(10deg, 45deg)] from accumulate [skew(20deg, 30deg)] to accumulate [skew(40deg, 70deg)] at (0.75) should be [skew(45deg, 105deg)] assert_equals: expected "matrix ( 1 , - 3.73 , 1 , 1 , 0 , 0 ) " but got "matrix ( 1 , 1.73 , 0.7 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skew(10deg, 45deg)] from accumulate [skew(20deg, 30deg)] to accumulate [skew(40deg, 70deg)] at (1) should be [skew(50deg, 115deg)] assert_equals: expected "matrix ( 1 , - 2.14 , 1.19 , 1 , 0 , 0 ) " but got "matrix ( 1 , 2.75 , 0.84 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skew(10deg, 45deg)] from accumulate [skew(20deg, 30deg)] to accumulate [skew(40deg, 70deg)] at (1.5) should be [skew(60deg, 135deg)] assert_equals: expected "matrix ( 1 , - 1 , 1.73 , 1 , 0 , 0 ) " but got "matrix ( 1 , 16331239353195370 , 1.19 , 1 , 0 , 0 ) "
FAIL Compositing: property <transform> underlying [skewX(45deg)] from accumulate [skewY(45deg)] to accumulate [skewY(45deg)] at (0.5) should be [matrix(1, 1, 0.5, 1.5, 0, 0)] assert_equals: expected "matrix ( 1 , 1 , 0.5 , 1.5 , 0 , 0 ) " but got "matrix ( 1 , 1 , 0 , 1 , 0 , 0 ) "
